Add agent metrics impl
